removed repeated username, multiple styling improvements
When someone posts multiple messages "in a row" to the list, and their nick
shows up for each line, the nick no longer shows up in the log multiple
times --- instead, just once for the first in the string of messages.

Also got rid of table lines, narrowed the window a bit, and changed
all the colors to use fewer of them, and darker ones at that.